requirements (1).txt
pillow
streamlit 
tensorflow 
nltk 
pandas>=2.0.3 
gensim 
scikit-learn
numpy>=1.24.3
scipy
gensim
